LPS-176939 Extract to BaseConfigurationModelListener avoiding duplicate code.
